requirements.txt
Flask
psycopg2-binary
SQLAlchemy
python-dotenv
openai
